+++ packages/thirdweb/src/adapters/ethers6.test.ts
@@ -1,37 +1,43 @@
 import { describe, test, expect } from "vitest";
-import * as ethers6 from "ethers6";
 import { privateKeyAccount } from "../wallets/private-key.js";
 import { TEST_CLIENT } from "../../test/src/test-clients.js";
-import { toEthersSigner } from "./ethers6.js";
-import { defineChain } from "../chains/utils.js";
-
-const FAKE_PKEY =
-  "0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef";
+import { ethers6Adapter } from "./ethers6.js";
+import { ANVIL_CHAIN } from "../../test/src/chains.js";
+import { ANVIL_PKEY_A, TEST_ACCOUNT_B } from "../../test/src/test-wallets.js";
 
 const account = privateKeyAccount({
-  privateKey: FAKE_PKEY,
+  privateKey: ANVIL_PKEY_A,
   client: TEST_CLIENT,
 });
 
 describe("toEthersSigner", () => {
   test("should return an ethers 6 signer", async () => {
-    const signer = await toEthersSigner(
-      ethers6,
-      TEST_CLIENT,
+    const signer = await ethers6Adapter.signer.toEthers({
+      client: TEST_CLIENT,
       account,
-      defineChain(31337),
-    );
+      chain: ANVIL_CHAIN,
+    });
     expect(signer).toBeDefined();
     expect(signer.signMessage).toBeDefined();
   });
 
-  test("should sign typed data", async () => {
-    const signer = await toEthersSigner(
-      ethers6,
-      TEST_CLIENT,
+  test("should sign message", async () => {
+    const signer = await ethers6Adapter.signer.toEthers({
+      client: TEST_CLIENT,
       account,
-      defineChain(31337),
-    );
+      chain: ANVIL_CHAIN,
+    });
+    const expectedSig = await account.signMessage({ message: "hello world" });
+    const sig = await signer.signMessage("hello world");
+    expect(sig).toBe(expectedSig);
+  });
+
+  test("should sign typed data", async () => {
+    const signer = await ethers6Adapter.signer.toEthers({
+      client: TEST_CLIENT,
+      account,
+      chain: ANVIL_CHAIN,
+    });
     expect(signer.signTypedData).toBeDefined();
 
     // All properties on a domain are optional
@@ -70,8 +76,19 @@ describe("toEthersSigner", () => {
 
     const signature = await signer.signTypedData(domain, types, value);
 
-    expect(signature).toMatchInlineSnapshot(
-      `"0x10d3ce8040590e48889801080ad40f3d514c2c3ce03bbbe3e179bbf5ba56c75425951fa15220f637e2ab79fd033b99c4b340339e00e360316547e956c61ffcb01c"`,
-    );
+    expect(signature.length).toBe(132);
+  });
+
+  test("should send a tx", async () => {
+    const signer = await ethers6Adapter.signer.toEthers({
+      client: TEST_CLIENT,
+      account,
+      chain: ANVIL_CHAIN,
+    });
+    const txResponse = await signer.sendTransaction({
+      to: TEST_ACCOUNT_B.address,
+      value: 100,
+    });
+    expect(txResponse.hash.length).toBe(66);
   });
 });