+++ packages/thirdweb/src/exports/react.ts
@@ -51,8 +51,9 @@ export {
 } from "../react/core/hooks/wallets/wallet-hooks.js";
 
 // contract related
-export { useReadContract } from "../react/core/hooks/contract/useRead.js";
-export { useSendTransaction } from "../react/core/hooks/contract/useSend.js";
+export { useReadContract } from "../react/core/hooks/contract/useReadContract.js";
+export { useSendTransaction } from "../react/core/hooks/contract/useSendTransaction.js";
+export { useSendAndConfirmTransaction } from "../react/core/hooks/contract/useSendAndConfirmTransaction.js";
 export { useEstimateGas } from "../react/core/hooks/contract/useEstimateGas.js";
 export { useEstimateGasCost } from "../react/core/hooks/contract/useEstimateGasCost.js";
 export { useWaitForReceipt } from "../react/core/hooks/contract/useWaitForReceipt.js";