+++ legacy_packages/react-native-compat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @thirdweb-dev/react-native-compat
 
+## 0.7.28
+
+## 0.7.27
+
 ## 0.7.26
 
 ### Patch Changes