+++ legacy_packages/wallets/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @thirdweb-dev/wallets
 
+## 2.4.33
+
+### Patch Changes
+
+- Updated dependencies [[`b3fca95`](https://github.com/thirdweb-dev/js/commit/b3fca9509feb0ac6e0ded816f08570b2954f6521)]:
+  - @thirdweb-dev/chains@0.1.88
+  - @thirdweb-dev/sdk@4.0.57
+
+## 2.4.32
+
+### Patch Changes
+
+- Updated dependencies [[`9af8561`](https://github.com/thirdweb-dev/js/commit/9af85612845573b54b54cb72e55404aad2c74700), [`fbf4556`](https://github.com/thirdweb-dev/js/commit/fbf4556d489b3be7c0c0ff85736e803573f55228)]:
+  - @thirdweb-dev/sdk@4.0.56
+  - @thirdweb-dev/chains@0.1.87
+
 ## 2.4.31
 
 ### Patch Changes